==> 00-determine_baseline.ipynb <==
#select with baseline to use for the z-scoring
#not a part of the main analysis

import preprocessing.data_preprocessing.refractor_utils as ru
import sys

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import scipy
from pathlib import Path

import utils.helper_func as hf
import utils.preprocessing as preprocessing
import utils.plotting as plotting

top_dir = '/home/anastasia/epiphyte/anastasia/output'
folder = '05-Hilbert_transform'

patient_id = 46
session_nr = 1

df_patient_info = pd.read_csv(f'{top_dir}/{patient_id}_channel_info.csv')

fs = 32768
#The resulting sample rate is up / down times the original sample rate.
up = 1
down=32
fs_downs = (up/down)*fs
dt = 1/fs_downs      # sampling period/time/interval or time resolution, often denoted as T

signal_len = np.int(fs_downs*1.5)+1
times = np.linspace(-0.5, 1, signal_len)
times = times*1000
time_zero_idx = np.where(times == hf.find_nearest(times, 0))[0][0]

df_stim_info = pd.read_csv(f'{top_dir}/{patient_id}_df_stim_info.csv')

all_stim = np.unique(df_stim_info['stim_id'])
all_stim_name = np.unique(df_stim_info['stim_name'])
all_brain_areas = np.unique(df_patient_info['recording_site'])

#filtering_types = ['slow_gamma']
filtering_types = ['theta', 'slow_gamma', 'fast_gamma']

x_pos = np.arange(len(filtering_types))

for i in range(len(df_patient_info['channel_name'])):
    ch = df_patient_info.loc[i,'channel_name']
    ch_site = df_patient_info.loc[i,'recording_site']
    
    baseline_500_pre = []
    baseline_100_pre = []
    baseline_500_post = []
    baseline_100_post = []
    std_500_pre = []
    std_100_pre = []
    std_500_post = []
    std_100_post = []
    
    for filtering_type in filtering_types:
        bsl_500_pre = np.load(f'{top_dir}/{folder}/{filtering_type}/baseline/{ch}_{ch_site}_mean_baseline_500_pre.npy')
        bsl_100_pre = np.load(f'{top_dir}/{folder}/{filtering_type}/baseline/{ch}_{ch_site}_mean_baseline_100_pre.npy')
        bsl_500_post = np.load(f'{top_dir}/{folder}/{filtering_type}/baseline/{ch}_{ch_site}_mean_baseline_500_post.npy')
        bsl_100_post = np.load(f'{top_dir}/{folder}/{filtering_type}/baseline/{ch}_{ch_site}_mean_baseline_100_post.npy')
        
        baseline_500_pre.append(bsl_500_pre)
        baseline_100_pre.append(bsl_100_pre)
        baseline_500_post.append(bsl_500_post)
        baseline_100_post.append(bsl_100_post)
        
        std_500_pre.append(np.std(bsl_500_pre))
        std_100_pre.append(np.std(bsl_100_pre))
        std_500_post.append(np.std(bsl_500_post))
        std_100_post.append(np.std(bsl_100_post))
    
    #baselines_pre = [aluminum_mean, copper_mean, steel_mean]
    #stds_pre = [aluminum_std, copper_std, steel_std]
    fig, (ax1, ax2, ax3, ax4) = plt.subplots(1, 4, sharey=True, figsize=(20, 5))
    title = f'Baseline, channel {ch}, recording_site {ch_site}'
    #file_name = f"{{'amplitude_envelope_and_mean_signal_channel'!'{ch}', 'recording_site'!'{ch_site}', 'stim_index'!'{current_stim_index}', 'stim_name'!'{current_stim_name}'}}"
    fig.suptitle(title)

    ax1.bar(x_pos, np.mean(baseline_500_pre,axis=1), yerr=std_500_pre, align='center', alpha=0.5, ecolor='black', capsize=10)
    ax1.set_xticks(x_pos)
    ax1.set_xticklabels(filtering_types)
    ax1.set_title('Baseline pre-movie 500 ms')
    ax1.set_ylabel('log10 power')
    ax2.bar(x_pos, np.mean(baseline_100_pre,axis=1), yerr=std_100_pre, align='center', alpha=0.5, ecolor='black', capsize=10)
    ax2.set_xticks(x_pos)
    ax2.set_xticklabels(filtering_types)
    ax2.set_title('Baseline pre-movie 100 ms')
    ax3.bar(x_pos, np.mean(baseline_500_pre,axis=1), yerr=std_500_post, align='center', alpha=0.5, ecolor='black', capsize=10)
    ax3.set_xticks(x_pos)
    ax3.set_xticklabels(filtering_types)
    ax3.set_title('Baseline post-movie 500 ms')
    ax4.bar(x_pos, np.mean(baseline_100_post,axis=1), yerr=std_100_post, align='center', alpha=0.5, ecolor='black', capsize=10)
    ax4.set_xticks(x_pos)
    ax4.set_xticklabels(filtering_types)
    ax4.set_title('Baseline post-movie 100 ms')
    
    #plt.show()
    fig.savefig(f'{top_dir}/{folder}/plots/baseline/{title}', facecolor='white', transparent=False)
    plt.close()

    
        


wilcoxon_post

x_pos = np.arange(4)
wilcoxon_pre_all = []
wilcoxon_post_all = []


for i in range(len(df_patient_info['channel_name'])):
    ch = df_patient_info.loc[i,'channel_name']
    print(ch)
    ch_site = df_patient_info.loc[i,'recording_site']
    
    baseline_500_pre = []
    baseline_100_pre = []
    baseline_500_post = []
    baseline_100_post = []
    std_500_pre = []
    std_100_pre = []
    std_500_post = []
    std_100_post = []
    baseline = []
    stds = []
    wilcoxon_pre = []
    wilcoxon_post = []
    
    for filtering_type in filtering_types:
        print(filtering_type)
        bsl_500_pre = np.load(f'{top_dir}/{folder}/{filtering_type}/baseline/{ch}_{ch_site}_mean_baseline_500_pre.npy')
        bsl_100_pre = np.load(f'{top_dir}/{folder}/{filtering_type}/baseline/{ch}_{ch_site}_mean_baseline_100_pre.npy')
        _, wilc_pre = scipy.stats.wilcoxon(bsl_500_pre, bsl_100_pre)
        print(wilc_pre)
        wilcoxon_pre.append(wilc_pre)
        bsl_500_post = np.load(f'{top_dir}/{folder}/{filtering_type}/baseline/{ch}_{ch_site}_mean_baseline_500_post.npy')
        bsl_100_post = np.load(f'{top_dir}/{folder}/{filtering_type}/baseline/{ch}_{ch_site}_mean_baseline_100_post.npy')
        _, wilc_post = scipy.stats.wilcoxon(bsl_500_post, bsl_100_post)
        print(wilc_post)
        wilcoxon_post.append(wilc_post)
        
        baseline.append(bsl_500_pre)
        baseline.append(bsl_100_pre)
        baseline.append(bsl_500_post)
        baseline.append(bsl_100_post)
        
        stds.append(np.std(bsl_500_pre))
        stds.append(np.std(bsl_100_pre))
        stds.append(np.std(bsl_500_post))
        stds.append(np.std(bsl_100_post))
    
    baselines_theta = baseline[0:4]
    baselines_slow_gamma = baseline[4:8]
    baselines_fast_gamma = baseline[8:12]
    std_theta = stds[0:4]
    std_slow_gamma = stds[4:8]
    std_fast_gamma = stds[8:12]
    
    wilcoxon_pre_all.append(wilcoxon_pre)
    wilcoxon_post_all.append(wilcoxon_post)
    
    xlabels = ['pre 500 ms', 'pre 100 ms', 'post 500 ms', 'post 100 ms', ]
    
    fig, (ax1, ax2, ax3) = plt.subplots(1, 3, sharey=True, figsize=(16, 5))
    title = f'Baseline, channel {ch}, recording_site {ch_site}'
    #file_name = f"{{'amplitude_envelope_and_mean_signal_channel'!'{ch}', 'recording_site'!'{ch_site}', 'stim_index'!'{current_stim_index}', 'stim_name'!'{current_stim_name}'}}"
    fig.suptitle(title)
    
    ax1.bar(x_pos, np.mean(baselines_theta,axis=1), yerr=std_theta, align='center', alpha=0.5, ecolor='black', capsize=10)
    ax1.set_xticks(x_pos)
    ax1.set_xticklabels(xlabels)
    ax1.set_title(f'theta, wilcoxon_pre = {wilcoxon_pre[0]}, wilcoxon_post = {wilcoxon_post[0]}')
    ax1.set_ylabel('log10 power')
    ax2.bar(x_pos, np.mean(baselines_slow_gamma,axis=1), yerr=std_slow_gamma, align='center', alpha=0.5, ecolor='black', capsize=10)
    ax2.set_xticks(x_pos)
    ax2.set_xticklabels(xlabels)
    ax2.set_title('slow gamma')
    ax3.bar(x_pos, np.mean(baselines_fast_gamma,axis=1), yerr=std_fast_gamma, align='center', alpha=0.5, ecolor='black', capsize=10)
    ax3.set_xticks(x_pos)
    ax3.set_xticklabels(xlabels)
    ax3.set_title('fast gamma')

    #plt.show()
    fig.savefig(f'{top_dir}/{folder}/plots/baseline_per_freq/{title}', facecolor='white', transparent=False)
    plt.close()

    
        


a = np.array(wilcoxon_post_all)
b = np.argwhere(a <= 0.05)
len(b)

np.arange(0,80,10)

import statsmodels
for i in np.arange(0,80,10):
    c = statsmodels.stats.multitest.multipletests(a[:i+10,2], alpha=0.05, method='simes-hochberg')[1]
    b = np.argwhere(c <= 0.05)
    print(len(b))

c

a
